Better, explicit capability check for document.createEvent
Don't assume that window.addEventListener capability gives us the
document.createEvent capability.

(Some IE shims, like aight.js, only define the former and not the
latter. In this case, timesheets.js assumes that the browser is
modern and then encounters a javascript error when it tries to
execute document.createEvent.)